Add files via upload
In this project, Codecademy provides a design spec and image assets to help  along the way. A design spec is an image of a web page outlined with all of its CSS properties and values.  Throughout the project, it is used with almost all kind of CSS properties such as box model, display and positioning, color, typography, responsive design like media query and flexbox.