Create app.js
In this project, I used JavaScript to randomly create a three-course meal based on what is available on a menu. As a frequent diner, I try out new restaurants and experiment with different foods. However, having to figure out what  want to order can be a time-consuming ordeal if the menu is big, and  want an easier way to be able to figure out what the customers are going to eat. In this project, I use the knowledge of JavaScript object and accessing the the property , method, getter ,setter function and factory function.