Add files via upload
Books-‘N-Stuff carries three different types of media: books, CDs, and movies. In this project I create a parent class named Media with three subclasses: Book, Movie, and CD. The three subclasses have their properties and methods In this project, I have used the knowledge of JavaScript's Class  which is a template literal of Object.