Create bicycle-world-ii.txt
Welcome to Bicycle World, the premier text-based bicycle shop! This shop is only accessible to programmers , who are familiar with the command line.

In this project, I have used the commands learned to navigate and edit the filesystem of this special shop.

The starting filesystem is shown above. (Bicycle World recently changed owners, who named the main directory bicycle-world-ii.)